feat/ref(*): auto-detect provider
